Rename folders and files containing new observables
naming is hard, suggestions welcome
