+++ packages/sdk/src/index.ts
@@ -50,10 +50,10 @@ export * from './streamStateView_UserInbox'
 export * from './streamStateView_UserMetadata'
 export * from './streamStateView_UserSettings'
 export * from './streamUtils'
-export * from './streams-view/streamsView'
-export * from './streams-view/timelinesView'
-export * from './streams-view/timelineEvents'
-export * from './streams-view/timelinesViewModel'
+export * from './views/streamsView'
+export * from './views/streams/timelines'
+export * from './views/streams/timelineEvents'
+export * from './views/streams/timelinesModel'
 export * from './sync-agent/db'
 export * from './sync-agent/dms/dms'
 export * from './sync-agent/dms/models/dm'
