+++ packages/sdk/src/tests/multi_ne/syncedStreams.test.ts
@@ -16,7 +16,7 @@ import { SyncedStream } from '../../syncedStream'
 import { StubPersistenceStore } from '../../persistenceStore'
 import { Envelope, StreamEvent, PlainMessage } from '@towns-protocol/proto'
 import { nanoid } from 'nanoid'
-import { StreamsView, StreamsViewDelegate } from '../../streams-view/streamsView'
+import { StreamsView, StreamsViewDelegate } from '../../views/streamsView'
 
 const log = dlog('csb:test:syncedStreams')
